Moves extension code, adds CMake
Moved extesion code out of where the sqf and mission files are so that that
whole area doesn't get cluttered. Extension code now in the Extensions
directory.

Also added a bootstrap script for getting cmake and some other stuff, look at
the instructions in README.md if you want to build it.